moves event gen scripts to own file
